+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - `@php`,`@component` to the list of directives that are shifted to start of line by @surgiie in https://github.com/surgiie/blade/pulls/1
 - `@endcomponent` gets a new line implictly added to avoid merging with next line on files by @surgiie in https://github.com/surgiie/blade/pulls/1
 - `<x-*>` blade component support by @surgiie in https://github.com/surgiie/blade/pulls/1
+- Flush file finder after compile to ensure cached view property doesnt return wrong file contents @surgiie in https://github.com/surgiie/blade/pulls/1
+- Add `setInstance` and `getInstance` static methods on `Blade` by @surgiie in https://github.com/surgiie/blade/pulls/1
+- `blade()` helper that returns static instance by @surgiie in https://github.com/surgiie/blade/pulls/1
 
 ## [v0.1.0] - 2022-10-20